Prevent errors when swapping very large amounts of tokens on the platform
Adds overflow checks and safe conversions to BigInt in swap calculations to prevent "value out of range" errors.

+++ client/src/pages/FixedSwapPage.tsx
@@ -198,13 +198,14 @@ export default function FixedSwapPage() {
         
         // CRITICAL FIX: Convert the input amount to raw blockchain format (lamports)
         // This ensures the amount shown in UI is exactly what's sent to the blockchain
+        // Use BigInt for large numbers to prevent numeric overflow
         const rawFromAmount = fromToken === SOL_SYMBOL 
-          ? Math.floor(fromAmount * 1e9) // Convert SOL to lamports 
-          : Math.floor(fromAmount * 1e9); // Convert YOT to raw amount with 9 decimals
+          ? BigInt(Math.floor(fromAmount * 1e9)) // Convert SOL to lamports 
+          : BigInt(Math.floor(fromAmount * 1e9)); // Convert YOT to raw amount with 9 decimals
           
         const rawToAmount = toToken === SOL_SYMBOL
-          ? Math.floor(toAmount * 1e9) // Convert SOL to lamports
-          : Math.floor(toAmount * 1e9); // Convert YOT to raw amount with 9 decimals
+          ? BigInt(Math.floor(toAmount * 1e9)) // Convert SOL to lamports
+          : BigInt(Math.floor(toAmount * 1e9)); // Convert YOT to raw amount with 9 decimals
           
         console.log(`Converted raw amounts: ${rawFromAmount} ${fromToken} -> ${rawToAmount} ${toToken}`);
